+++ dscript/__init__.py
@@ -1,17 +1,23 @@
 __version__ = "0.2.1-dev"
-__citation__ = "Sledzieski, Singh, Cowen, Berger. Sequence-based prediction of protein-protein interactions: a structure-aware interpretable deep learning model. Under Review. 2020."
-import dscript.alphabets
-import dscript.commands
-import dscript.language_model
-import dscript.models
-import dscript.pretrained
-import dscript.utils
+__citation__ = """Sledzieski, Singh, Cowen, Berger. "D-SCRIPT translates genome to phenome with sequence-based, structure-aware, genome-scale predictions of protein-protein interactions." Cell Systems 12, no. 10 (2021): 969-982.
+
+Devkota, Singh, Sledzieski, Berger, Cowen, Topsy-Turvy: integrating a global view into sequence-based PPI prediction, Bioinformatics, In Press."""
+from . import (
+    alphabets,
+    commands,
+    fasta,
+    glider,
+    language_model,
+    models,
+    pretrained,
+)
 
 __all__ = [
     "models",
     "commands",
     "alphabets",
     "fasta",
+    "glider",
     "language_model",
     "pretrained",
     "utils",